+++ scripts/measures/build-measures
@@ -67,12 +67,7 @@ echo "validating new measures-data.json against schema"
 cat measures/$currentPerformanceYear/measures-data.json | \
 	node scripts/validate-data.js measures $currentPerformanceYear
 
-# 4. Validate that firstPerformanceYear values match actual measures in that prior year
-# We don't have measures data from prior to 2017, so don't do this for
-# performance years earlier than 2018
-if [ $currentPerformanceYear -gt 2017 ]
-then
-	echo "validating that firstPerformanceYear values match actual measures in that prior year"
-	cat measures/$currentPerformanceYear/measures-data.json | \
-		node scripts/validate-first-performance-year.js $currentPerformanceYear
-fi
+# 4. Validate that measures actually existed in their listed firstPerformanceYear and the years after
+echo "validating that measures actually existed in their listed firstPerformanceYear and the years after "
+cat measures/$currentPerformanceYear/measures-data.json | \
+	node scripts/validate-measures-past-existence.js $currentPerformanceYear
